feat: move Method Types to mod
