+++ substrate/client/rpc-spec-v2/src/archive/archive.rs
@@ -21,6 +21,7 @@
 use crate::{
 	archive::{
 		archive_storage::ArchiveStorageDiff, error::Error as ArchiveError, ArchiveApiServer,
+		MethodResult,
 	},
 	common::{
 		events::{
@@ -28,7 +29,7 @@ use crate::{
 		},
 		storage::{QueryResult, StorageSubscriptionClient},
 	},
-	hex_string, MethodResult, SubscriptionTaskExecutor,
+	hex_string, SubscriptionTaskExecutor,
 };
 
 use codec::Encode;
