+++ substrate/client/rpc-spec-v2/src/lib.rs
@@ -36,78 +36,7 @@ pub mod transaction;
 /// Task executor that is being used by RPC subscriptions.
 pub type SubscriptionTaskExecutor = std::sync::Arc<dyn sp_core::traits::SpawnNamed>;
 
-/// The result of an RPC method.
-#[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
-#[serde(untagged)]
-pub enum MethodResult {
-	/// Method generated a result.
-	Ok(MethodResultOk),
-	/// Method encountered an error.
-	Err(MethodResultErr),
-}
-
-impl MethodResult {
-	/// Constructs a successful result.
-	pub fn ok(result: impl Into<String>) -> MethodResult {
-		MethodResult::Ok(MethodResultOk { success: true, value: result.into() })
-	}
-
-	/// Constructs an error result.
-	pub fn err(error: impl Into<String>) -> MethodResult {
-		MethodResult::Err(MethodResultErr { success: false, error: error.into() })
-	}
-}
-
-/// The successful result of an RPC method.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct MethodResultOk {
-	/// Method was successful.
-	success: bool,
-	/// The result of the method.
-	pub value: String,
-}
-
-/// The error result of an RPC method.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct MethodResultErr {
-	/// Method encountered an error.
-	success: bool,
-	/// The error of the method.
-	pub error: String,
-}
-
 /// Util function to encode a value as a hex string
 pub fn hex_string<Data: AsBytesRef>(data: &Data) -> String {
 	format!("0x{:?}", HexDisplay::from(data))
-}
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-
-	#[test]
-	fn method_result_ok() {
-		let ok = MethodResult::ok("hello");
-
-		let ser = serde_json::to_string(&ok).unwrap();
-		let exp = r#"{"success":true,"value":"hello"}"#;
-		assert_eq!(ser, exp);
-
-		let ok_dec: MethodResult = serde_json::from_str(exp).unwrap();
-		assert_eq!(ok_dec, ok);
-	}
-
-	#[test]
-	fn method_result_error() {
-		let ok = MethodResult::err("hello");
-
-		let ser = serde_json::to_string(&ok).unwrap();
-		let exp = r#"{"success":false,"error":"hello"}"#;
-		assert_eq!(ser, exp);
-
-		let ok_dec: MethodResult = serde_json::from_str(exp).unwrap();
-		assert_eq!(ok_dec, ok);
-	}
 }
